# app/Http/Controllers/AuthController.php
<?php

namespace App\Http\Controllers;

use Illuminate\Http\Request;
use Illuminate\Http\Response;

use App\Models\User;
use App\Models\Admin;

use App\Helpers\Validation;
use Illuminate\Support\Facades\Hash;

/**
 * @OA\Info(
 *     title="Custom Analytic - WW2 API",
 *     version="1.0.0",
 *     description="API Documentation for Custom Analytic - WW2",
 *     @OA\Contact(
 *         email="flazen.edu@gmail.com"
 *     )
 * )
*/

class AuthController extends Controller
{
    /**
     * @OA\POST(
     *     path="/api/login",
     *     summary="Sign in to the Apps",
     *     tags={"Auth"},
     *     @OA\Response(
     *         response=200,
     *         description="{user_data}"
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="{validation_msg}"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Wrong username or password"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Internal Server Error"
     *     ),
     * )
     */
    public function login(Request $request)
    {
        try {
            $validator = Validation::getValidateLogin($request);

            if ($validator->fails()) {
                $errors = $validator->messages();

                return response()->json([
                    'status' => 'failed',
                    'message' => $errors,
                    'token' => null
                ], Response::HTTP_UNPROCESSABLE_ENTITY);
            } else {
                $role = 'admin';
                $user = Admin::where('username', $request->username)->first();

                if($user == null){
                    $user = User::where('username', $request->username)->first();
                    $role = 'user';
                } 

                if (!$user || !Hash::check($request->password, $user->password)) {
                    return response()->json([
                        'status' => 'failed',
                        'message' => 'Wrong username or password',
                        'token' => null
                    ], Response::HTTP_UNAUTHORIZED);
                } else {
                    $token = $user->createToken('login')->plainTextToken;

                    return response()->json([
                        'status' => 'success',
                        'result' => $user,
                        'token' => $token,
                        'role' => $role                 
                    ], Response::HTTP_OK);
                }
            } 
        } catch(\Exception $e) {
            return response()->json([
                'status' => 'error',
                'message' => 'something wrong. please contact admin',
            ], Response::HTTP_INTERNAL_SERVER_ERROR);
        }
    }

    /**
     * @OA\GET(
     *     path="/api/logout",
     *     summary="Sign out from Apps",
     *     tags={"Auth"},
     *     @OA\Response(
     *         response=200,
     *         description="Logout success"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Internal Server Error"
     *     ),
     * )
     */
    public function logout(Request $request)
    {
        try {
            $request->user()->currentAccessToken()->delete();

            return response()->json([
                'status' => 'success',
                'message' => 'Logout success'
            ], Response::HTTP_OK);
        } catch(\Exception $e) {
            return response()->json([
                'status' => 'error',
                'message' => 'something wrong. please contact admin',
            ], Response::HTTP_INTERNAL_SERVER_ERROR);
        }
    }

    /**
     * @OA\GET(
     *     path="/api/v1/user/my",
     *     summary="Get my profile info",
     *     tags={"Auth"},
     *     @OA\Response(
     *         response=200,
     *         description="User found"
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="User not found"
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Internal Server Error"
     *     ),
     * )
     */
    public function getMyProfile(Request $request){
        try{
            $user_id = $request->user()->id;

            $usr = User::select('fullname','username','email','profile_img','bio','email_verified_at','created_at','updated_at')
                ->where('id', $user_id)
                ->first();

            if($usr){
                return response()->json([
                    "message"=> "User found", 
                    "status"=> 'success',
                    "data"=> $usr
                ], Response::HTTP_OK);
            } else {
                $adm = Admin::select('fullname','username','bio','email','created_at','updated_at')
                    ->where('id', $user_id)
                    ->first();

                if($adm){
                    return response()->json([
                        "message"=> "Admin found", 
                        "status"=> 'success',
                        "data"=> $adm
                    ], Response::HTTP_OK);
                } else {
                    return response()->json([
                        "message"=> "User not found", 
                        "status"=> 'success',
                        "data"=> null
                    ], Response::HTTP_NOT_FOUND);
                }
            }
        } catch(\Exception $e) {
            return response()->json([
                'status' => 'error',
                'message' => 'something wrong. please contact admin'
            ], Response::HTTP_INTERNAL_SERVER_ERROR);
        }
    }
}