# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Custom Analytic - WW2 API",
        "description": "API Documentation for Custom Analytic - WW2",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in to the Apps",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "responses": {
                    "200": {
                        "description": "{user_data}"
                    },
                    "422": {
                        "description": "{validation_msg}"
                    },
                    "401": {
                        "description": "Wrong username or password"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign out from Apps",
                "operationId": "3f19b742e4ead13f81ba0d6c0d46ac1b",
                "responses": {
                    "200": {
                        "description": "Logout success"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/user/my": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get my profile info",
                "operationId": "7a7add02ec426511cfeb72fb4dfb1615",
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}